DT-835 Convert court report controller tests to unit tests
